+++ src/main/java/com/sehee/weeklyjpa/domain/customer/CustomerRepository.java
@@ -1,5 +1,6 @@
-package com.sehee.weeklyjpa.domain;
+package com.sehee.weeklyjpa.domain.customer;
 
+import com.sehee.weeklyjpa.domain.customer.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
